rag-flow/src/requirements.txt
sentence-transformers>=2.2.2
PyMuPDF>=1.21.1
faiss-cpu>=1.7.4
numpy>=1.24.0 